+++ sensing/preprocessor/gnss/gnss_poser/include/gnss_poser/gnss_poser_node.hpp
@@ -16,16 +16,16 @@
 #ifndef GNSS_POSER_CORE_HPP_
 #define GNSS_POSER_CORE_HPP_
 
-#include <gnss_poser/convert.hpp>
-#include <gnss_poser/gnss_stat.hpp>
+#include "gnss_poser/convert.hpp"
+#include "gnss_poser/gnss_stat.hpp"
 
-#include <rclcpp/rclcpp.hpp>
 #include <tf2/transform_datatypes.h>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 #include <tf2_ros/transform_broadcaster.h>
 #include <tf2_ros/transform_listener.h>
 #include <geometry_msgs/msg/pose_stamped.hpp>
 #include <geometry_msgs/msg/pose_with_covariance_stamped.hpp>
+#include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/nav_sat_fix.hpp>
 #include <std_msgs/msg/bool.hpp>
 #include <ublox_msgs/msg/nav_pvt.hpp>
@@ -35,7 +35,7 @@
 class GNSSPoser : public rclcpp::Node
 {
 public:
-  GNSSPoser();
+  explicit GNSSPoser(const rclcpp::NodeOptions & node_options);
 
 private:
   void callbackNavSatFix(const sensor_msgs::msg::NavSatFix::ConstSharedPtr nav_sat_fix_msg_ptr);
